improve text for tank sensor name
Use customname or devicename instead of Connection when available.
In case of multiple BLE sensors, it is impossible to select the correct
one when they are all called 'Fuel tank on Bluetooth LE'. If a
CustomName has been set, show that one, else fallback to the
DeviceName. If that fails too, fallback to the Mgmt/Connection.